apps/backend/services/pii-anonymization/requirements.txt
# PII Anonymization Service Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Security
PyJWT==2.8.0

# Regex and text processing
regex==2023.10.3

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# HTTP Client
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
PyYAML==6.0.1

